Updated devise registration create method to access payment if the plan is not free

+++ app/controllers/user/registrations_controller.rb
@@ -7,22 +7,23 @@ class User::RegistrationsController < Devise::RegistrationsController
       resource.save
       yield resource if block_given?
       if resource.persisted?
-        @payment = Payment.new({  email: params["user"]["email"],
-                                  token: params[:payment]["token"],
-                                  user_id: resource.id
-                              })
-        
-        flash[:error] = "Please check registration errors" unless @payment.valid?
-        begin
-          @payment.process_payment
-          @payment.save
-        rescue Exception => e
-          flash[:error] = e.message
-          resource.destroy
-          puts 'Payment failed'
-          render :new and return
+        unless resource.plan.name == "Free"
+          @payment = Payment.new({  email: params["user"]["email"],
+                                    token: params[:payment]["token"],
+                                    user_id: resource.id
+                                })
+          
+          flash[:error] = "Please check registration errors" unless @payment.valid?
+          begin
+            @payment.process_payment(resource.plan.price.to_i)
+            @payment.save
+          rescue Exception => e
+            flash[:error] = e.message
+            resource.destroy
+            puts 'Payment failed'
+            render :new and return
+          end
         end
-
         if resource.active_for_authentication?
           set_flash_message :notice, :signed_up if is_flashing_format?
           sign_up(resource_name, resource)
